feature: improve memory performances
This commit slightly improves memory performances by splitting the
process method into 2 steps (inference and alignment). This also
improves the way variables are assigned and try to avoid numpy array
duplication as much as possible. To do this and in order to keep the
original code of VDA unmodified, a new inherited class
"optimizedVideoDepthAnything" is defined.

While processing in chunks would be a more robust solution, implementing
it is challenging due to the sliding window mechanism used during
processing and the alignment step, which requires all generated depths.
A more comprehensive refactor could address this, but it would require
significantly more effort and time.